requirements.txt
amqp==2.2.2
billiard==3.5.0.3
cassandra-driver==3.12.0
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.7
django-analytical==2.2.2
django-cassandra-engine==1.3.0
django-extensions==1.9.7
django-filter==1.1.0
djangorestframework==3.3.3
geoip2==2.5.0
gunicorn==19.7.1
idna==2.6
kafka==1.3.5
kafka-python==1.3.5
kombu==4.1.0
Markdown==2.6.9
maxminddb==1.3.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
six==1.11.0
typing==3.6.2
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
wincertstore==0.2
